+++ OUDS/Core/Themes/Orange/Sources/Providers/ComponentTokens/OrangeThemeChipComponentTokensProvider.swift
@@ -77,7 +77,7 @@ import OUDSThemesContract
 ///     // - OrangeThemeDimensionSemanticTokensProvider for dimensions
 ///     let chipComponentTokensProvider = OrangeThemeChipComponentTokensProvider()
 ///
-///     // Or use your own _, border, color and space semantic tokens providers (or only some)
+///     // Or use your own dimension, border, color and space semantic tokens providers (or only some)
 ///     let chipComponentTokensProvider = OrangeThemeChipComponentTokensProvider(
 ///                                                 borders: CustomBorderSemanticTokensProvider(),
 ///                                                 colors: CustomColorSemanticTokensProvider(),
@@ -113,7 +113,7 @@ open class OrangeThemeChipComponentTokensProvider: AllChipComponentTokensProvide
     ///    - borders: Provider for border semantic tokens. If nil, a default one will be used (``OrangeThemeBorderSemanticTokensProvider``)
     ///    - colors: Provider for color semantic tokens. If nil, a default one will be used (``OrangeThemeColorSemanticTokensProvider``)
     ///    - spaces: Provider for space semantic tokens. If nil, a default one will be used (``OrangeThemeSpaceSemanticTokensProvider``)
-    ///    - dimensions: Provider for _ semantic tokens. If nil, a default one will be used (``OrangeThemeDimensionSemanticTokensProvider``)
+    ///    - dimensions: Provider for dimension semantic tokens. If nil, a default one will be used (``OrangeThemeDimensionSemanticTokensProvider``)
     public init(sizes: AllSizeSemanticTokensProvider? = nil,
                 borders: AllBorderSemanticTokensProvider? = nil,
                 colors: AllColorSemanticTokensProvider? = nil,