+++ OUDS/Core/Themes/Sosh/Sources/Providers/SemanticTokens/SoshThemeDimensionSemanticTokensProvider.swift
@@ -14,11 +14,11 @@
 import OUDSFoundations
 import OUDSThemesContract
 
-/// A class which wraps all **_ semantic tokens**, and expose them.
+/// A class which wraps all **dimension semantic tokens**, and expose them.
 /// This provider should be integrated as a `AllDimensionSemanticTokensProvider` implementation inside `OUDSTheme` so as to provide
 /// all tokens to the users.
 ///
-/// These _ semantic tokens are not overridable by design.
+/// These dimension semantic tokens are not overridable by design.
 /// They are considered as "closed" tokens but still defined in this library so as to keep consistancy between Figma specifications
 /// and library. The tokens are not hidden from developers.
 ///
@@ -47,6 +47,6 @@ final class SoshThemeDimensionSemanticTokensProvider: AllDimensionSemanticTokens
     // ଘ( ･ω･)_/ﾟ･:*:･｡☆
     // Note: So as to help the integration of generated code produced by the tokenator
     // the implemention of DimensionSemanticTokens is not here
-    // but in Core/Themes/Sosh/Values/DimensionTokens/SoshTheme+DimensionSemanticTokens.swift
+    // but in Core/Themes/Sosh/Values/SemanticTokens/SoshTheme+DimensionSemanticTokens.swift
     // This declaration of SoshThemeDimensionSemanticTokensProvider is here to allow to write documentation.
 }