+++ OUDS/Core/Themes/Wireframe/Sources/Providers/SemanticTokens/WireframeThemeDimensionSemanticTokensProvider.swift
@@ -16,11 +16,11 @@ import OUDSThemesContract
 
 // swiftlint:disable type_name
 
-/// A class which wraps all **_ semantic tokens**, and expose them.
+/// A class which wraps all **dimension semantic tokens**, and expose them.
 /// This provider should be integrated as a `AllDimensionSemanticTokensProvider` implementation inside `OUDSTheme` so as to provide
 /// all tokens to the users.
 ///
-/// These _ semantic tokens are not overridable by design.
+/// These dimension semantic tokens are not overridable by design.
 /// They are considered as "closed" tokens but still defined in this library so as to keep consistancy between Figma specifications
 /// and library. The tokens are not hidden from developers.
 ///
@@ -49,7 +49,7 @@ final class WireframeThemeDimensionSemanticTokensProvider: AllDimensionSemanticT
     // ଘ( ･ω･)_/ﾟ･:*:･｡☆
     // Note: So as to help the integration of generated code produced by the tokenator
     // the implemention of DimensionSemanticTokens is not here
-    // but in Core/Themes/Wireframe/Values/DimensionTokens/WireframeTheme+DimensionSemanticTokens.swift
+    // but in Core/Themes/Wireframe/Values/SemanticTokens/WireframeTheme+DimensionSemanticTokens.swift
     // This declaration of WireframeThemeDimensionSemanticTokensProvider is here to allow to write documentation.
 }