+++ OUDS/Core/ThemesContract/Sources/OUDSTheme/OUDSTheme.swift
@@ -58,6 +58,9 @@ open class OUDSTheme: @unchecked Sendable {
     public let charts: AllColorChartSemanticTokensProvider!
     // swiftlint:enable implicitly_unwrapped_optional
 
+    /// All effect semantic tokens exposed in one object
+    public let effects: AllEffectSemanticTokensProvider
+
     /// All elevation semantic tokens exposed in one object
     public let elevations: AllElevationSemanticTokensProvider
 
@@ -84,6 +87,9 @@ open class OUDSTheme: @unchecked Sendable {
 
     // MARK: - Component tokens
     // Keep things alhabetically ordered
+
+    /// All components tokens related to bar components
+    public let bar: AllBarComponentTokensProvider
 
     /// All components tokens related to badge components like `OUDSBadge`
     public let badge: AllBadgeComponentTokensProvider
@@ -165,6 +171,7 @@ open class OUDSTheme: @unchecked Sendable {
     ///    - colors: All semantic tokens of colors
     ///    - colorModes: All semantic tokens of color modes
     ///    - charts: All semantic tokens of color charts if the theme have some, otherwise nil (by default is nil)
+    ///    - effects: All semantic tokens of effects
     ///    - elevations: All semantic tokens of elevations
     ///    - fonts: All semantic tokens of fonts
     ///    - grids: All semantic tokens of grids
@@ -173,6 +180,7 @@ open class OUDSTheme: @unchecked Sendable {
     ///    - sizes: All semantic tokens of sizes
     ///    - spaces: All semantic tokens of spaces
     ///    - badge: All component tokens for badge
+    ///    - bar: All component tokens for bar
     ///    - bulletList: All component tokens for bullet list
     ///    - button: All component tokens for button
     ///    - checkbox: All component tokens for checkbox
@@ -199,6 +207,7 @@ open class OUDSTheme: @unchecked Sendable {
                 colors: AllColorSemanticTokensProvider,
                 colorModes: AllColorModeSemanticTokensProvider,
                 charts: AllColorChartSemanticTokensProvider? = nil,
+                effects: AllEffectSemanticTokensProvider,
                 elevations: AllElevationSemanticTokensProvider,
                 fonts: AllFontSemanticTokensProvider,
                 grids: AllGridSemanticTokensProvider,
@@ -207,6 +216,7 @@ open class OUDSTheme: @unchecked Sendable {
                 sizes: AllSizeSemanticTokensProvider,
                 spaces: AllSpaceSemanticTokensProvider,
                 badge: AllBadgeComponentTokensProvider,
+                bar: AllBarComponentTokensProvider,
                 bulletList: AllBulletListComponentTokensProvider,
                 button: AllButtonComponentTokensProvider,
                 checkbox: AllCheckboxComponentTokensProvider,
@@ -237,6 +247,7 @@ open class OUDSTheme: @unchecked Sendable {
         self.colorModes = colorModes
         self.charts = charts
         self.elevations = elevations
+        self.effects = effects
         self.fonts = fonts
         self.grids = grids
         self.opacities = opacities
@@ -246,6 +257,7 @@ open class OUDSTheme: @unchecked Sendable {
 
         // Save component tokens providers
         self.badge = badge
+        self.bar = bar
         self.button = button
         self.bulletList = bulletList
         self.checkbox = checkbox